backend/requirements.txt
fastapi
uvicorn
yfinance
httpx
psycopg2-binary
beautifulsoup4
textblob
nltk
pandas
scipy
apscheduler
sqlalchemy
python-dotenv
litellm
# transformers (Commented out for Railway Free Tier - 8GB limit)
# torch --index-url https://download.pytorch.org/whl/cpu (Commented out, too large)
